requirements.txt
flask
openai
python-dotenv
PyPDF2
python-docx
reportlab
SpeechRecognition
pydub
